feat(#40): Base of animation framework
Add refs to all the balls in order to be able to animate them efficiently.
It's important to animate the ref and not call setState() for performance.
Updating refs outside of the useFrame loop should be good enough.

This is the base for animations, we now need the array with moves from the backend

closes #40